pyproject.toml
[project]
name = "blog_generator_ai_agent"
version = "0.1.0"
description = "Blog-generator-ai-agent using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "chromadb>=0.5.23",
    "crewai[tools]>=0.159.0,<1.0.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "firecrawl-py>=4.1.1",
    "google-generativeai>=0.8.0,<0.9.0",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.0,<2.1.0",
    "loguru>=0.7.3",
    "markdown>=3.8.2",
    "mypy>=1.17.1",
    "nltk>=3.9.1",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.9.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "readability>=0.3.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.1",
    "uvicorn[standard]>=0.35.0",
    "yake>=0.6.0",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp-proto-grpc",
    "azure-monitor-opentelemetry",
    "azure-monitor-opentelemetry-exporter",
    "opentelemetry-instrumentation-fastapi"
]

[project.scripts]
blog_generator_ai_agent = "blog_generator_ai_agent.main:run"
run_crew = "blog_generator_ai_agent.main:run"
train = "blog_generator_ai_agent.main:train"
replay = "blog_generator_ai_agent.main:replay"
test = "blog_generator_ai_agent.main:test"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
